replace serial with identity, fix comments

+++ migrations/000004_create_book_categories_table.up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS book_categories (
-    book_id INT NOT NULL REFERENCES books(id) ON DELETE CASCADE,
-    category_id INT NOT NULL REFERENCES categories(id) ON DELETE CASCADE,
+    book_id BIGINT NOT NULL REFERENCES books(id) ON DELETE CASCADE,
+    category_id BIGINT NOT NULL REFERENCES categories(id) ON DELETE CASCADE,
     PRIMARY KEY (book_id, category_id)
 );
